Clean up
I should probably be using comprehensions all the time instead of
filter() and map()?

I wonder if this is an okay pattern:
if ...:
    ...
elif ...:
    ...
else:
    raise AssertionError("You shouldn't see this")

Found an example of that here:
https://github.com/rdkit/mmpdb/blob/master/mmpdblib/fragment_algorithm.py

Also, which is better?
re.match(r'...$', ...)
re.search(r'^...$', ...)
I think I prefer the latter because it's more explicit.